\name{contingency.table}
\alias{contingency.table}
\title{Contingency table.}
\usage{contingency.table(gene.set, p.val, sign=0.05)}
\description{Contingency table. Necessary for Fisher test.}
\author{Stephan Artmann}
\arguments{\item{gene.set}{Vector of gene sets.}
\item{p.val}{Vector with p-values.}
\item{sign}{Significane threshold.}}
