\name{fisher.combination}
\alias{fisher.combination}
\title{Fisher method of p value combination.}
\usage{fisher.combination(p1, p2, check.range=FALSE)}
\description{Fisher method of p value combination.}
\value{Combined p-value.}
\author{Stephan Artmann}
\arguments{\item{p1,p2}{one-sided p-values that shall be combined.}
\item{check.range}{If set to "TRUE" values above 1 will be set to 1.}}
