\name{inverse.normal.combination}
\alias{inverse.normal.combination}
\title{Inverse-normal method for p value combination.}
\usage{inverse.normal.combination(p1, p2)}
\description{Inverse-normal method for p value combination.}
\value{Two-sided combined p-value.}
\author{Stephan Artmann}
\arguments{\item{p1,p2}{one-sided p-values that shall be combined.}}
