\name{limma.one.sided}
\alias{limma.one.sided}
\title{Internal algorithm: Make limma test one-sided...}
\usage{limma.one.sided(fit, lower=FALSE)}
\description{Internal algorithm: Make limma test one-sided}
\arguments{\item{fit}{Result of "lmFit" and "eBayes" functions in "limma" package.}
\item{lower}{Shall one-sided p-value indicated down-regultation?}}
