\name{limma.test}
\alias{limma.test}
\title{internal algorithm for author's convenience.}
\usage{limma.test(X, group, design)}
\description{internal algorithm for author's convenience. Create a linear model with the limma package.}
\author{Stephan Artmann}
\arguments{\item{X}{Expression matrix.}
\item{group}{Group membership of replicates.}
\item{design}{Design as specified in limma (design matrix, see model.matrix).}}
