\name{m.combine}
\alias{m.combine}
\title{Internal function for author's convenience and more legible code.}
\usage{m.combine(M, v, FUN, ...)}
\description{Internal function for author's convenience and more legible code. Applies a function to every column vector of a matrix and a vector.}
\author{Stephan Artmann}
\arguments{\item{M}{The matrix for whose column vectors mapply shall be used.}
\item{v}{The vector.}
\item{FUN}{The function.}
\item{...}{Further arguments to be given to FUN.}}
